# 6% perf gain: Don't write fields to strings preemptively