This seems to improve performance by 5.7%!

```
format event/colors, wrapping, stderr
    time:   [11.696 µs 11.714 µs 11.737 µs]
    change: [-6.2772% -5.7379% -5.1220%] (p = 0.00 < 0.05)
format event/colors, wrapping
    time:   [5.2662 µs 5.2711 µs 5.2747 µs]
    change: [-14.121% -13.943% -13.777%] (p = 0.00 < 0.05)
format event/no colors, wrapping
    time:   [5.1020 µs 5.1102 µs 5.1164 µs]
    change: [-10.160% -10.017% -9.8873%] (p = 0.00 < 0.05)
format event/colors, no wrapping
    time:   [1.7943 µs 1.7968 µs 1.8006 µs]
    change: [-25.048% -24.892% -24.716%] (p = 0.00 < 0.05)
format event/no colors, no wrapping
    time:   [1.4897 µs 1.4916 µs 1.4947 µs]
    change: [-22.072% -21.960% -21.832%] (p = 0.00 < 0.05)
```